Removed old Json strategy and replaced with the logic of JsonSession. Removed Yaml strategy since it is unneeded. Removed setex since it unsupported (originally came from Rails 3.2 branch). Fixed tests to correctly insert ASCII characters. Added additional logic to raw encode strings that are ASCII-8Bit.
