+++ lib/redis/store/strategy/json.rb
@@ -4,17 +4,28 @@ class Redis
   class Store < self
     module Strategy
       module Json
+
+        class Error < StandardError
+        end
+
+        class SerializationError < Redis::Store::Strategy::Json::Error
+          def initialize(object)
+            super "Cannot correctly serialize object: #{object.inspect}"
+          end
+        end
+
         private
-          SERIALIZABLE = [String, TrueClass, FalseClass, NilClass, Numeric, Date, Time].freeze
-          MARSHAL_INDICATORS = ["\x04", "\004", "\u0004"].freeze
+          SERIALIZABLE = [String, TrueClass, FalseClass, NilClass, Numeric, Date, Time]
+          MARSHAL_INDICATORS = ["\x04", "\004", "\u0004"]
 
           def _dump(object)
             object = _marshal(object)
-            object.to_json
+            JSON.generate(object)
           end
 
           def _load(string)
-            object = JSON.parse(string, :symbolize_names => true)
+            object =
+              string.start_with?(*MARSHAL_INDICATORS) ? ::Marshal.load(string) : JSON.parse(string, :symbolize_names => true)
             _unmarshal(object)
           end
 
@@ -24,10 +35,13 @@ class Redis
               object.each { |k,v| object[k] = _marshal(v) }
             when Array
               object.each_with_index { |v, i| object[i] = _marshal(v) }
+            when String
+              object = object.to_json_raw_object if object.encoding == Encoding::ASCII_8BIT
+              object
             when *SERIALIZABLE
               object
             else
-              ::Marshal.dump(object)
+              raise SerializationError.new(object)
             end
           end
 
@@ -38,7 +52,7 @@ class Redis
             when Array
               object.each_with_index { |v, i| object[i] = _unmarshal(v) }
             when String
-              object.start_with?(MARSHAL_INDICATORS) ? ::Marshal.load(object) : object
+              object.start_with?(*MARSHAL_INDICATORS) ? ::Marshal.load(object) : object
             else
               object
             end
