+++ spec/redis/store/strategy/json_spec.rb
@@ -2,8 +2,9 @@ require 'spec_helper'
 
 describe "Redis::Store::Strategy::Json" do
   before(:each) do
+    @marshal_store = Redis::Store.new :strategy => :marshal
     @store = Redis::Store.new :strategy => :json
-    @rabbit = OpenStruct.new :name => 'rabbit', :legs => 4
+    @rabbit = {:name => "rabbit", :legs => 4}
     @peter     = { :name => "Peter Cottontail",
                    :race => @rabbit }
     @bunnicula = { :name    => "Bunnicula",
@@ -29,12 +30,12 @@ describe "Redis::Store::Strategy::Json" do
   end
 
   it "doesn't unmarshal on get if raw option is true" do
-    race = Marshal.dump(@rabbit).to_json
+    race = @rabbit.to_json
     @store.get("rabbit", :raw => true).should eql(%({"name":"Bunnicula","race":#{race},"friends":[{"name":"Peter Cottontail","race":#{race}}],"age":3.1,"alive":true}))
   end
 
   it "doesn't marshal on set if raw option is true" do
-    race = Marshal.dump(@rabbit)
+    race = @rabbit
     @store.set "rabbit", @peter, :raw => true
     @store.get("rabbit", :raw => true).should eql(%({:name=>"Peter Cottontail", :race=>#{race.inspect}}))
   end
@@ -51,7 +52,7 @@ describe "Redis::Store::Strategy::Json" do
 
   it "doesn't marshal on set unless exists if raw option is true" do
     @store.setnx "rabbit2", @peter, :raw => true
-    race = Marshal.dump(@rabbit)
+    race = @rabbit
     @store.get("rabbit2", :raw => true).should eql(%({:name=>"Peter Cottontail", :race=>#{race.inspect}}))
   end
 
@@ -66,31 +67,42 @@ describe "Redis::Store::Strategy::Json" do
     @store.set "rabbit", @bunnicula
     @store.set "rabbit2", @peter
     rabbit, rabbit2 = @store.mget "rabbit", "rabbit2", :raw => true
-    race = Marshal.dump(@rabbit).to_json
+    race = @rabbit.to_json
     rabbit.should eql(%({"name":"Bunnicula","race":#{race},"friends":[{"name":"Peter Cottontail","race":#{race}}],"age":3.1,"alive":true}))
     rabbit2.should eql(%({"name":"Peter Cottontail","race":#{race}}))
   end
 
-  describe "binary safety" do
+  it "will throw an error if an object isn't supported" do
+    lambda{
+      @store.set "rabbit2", OpenStruct.new(foo:'bar')
+    }.should raise_error Redis::Store::Strategy::Json::SerializationError
+  end
+
+  it "is able to bring out data that is marshalled using Ruby" do
+    @marshal_store.set "rabbit", @peter
+    rabbit = @store.get "rabbit"
+    rabbit.should eql(@peter)
+  end
+
+  context "binary safety" do
     before do
       @utf8_key = [51339].pack("U*")
       @ascii_string = [128].pack("C*")
-      @ascii_rabbit = OpenStruct.new(:name => @ascii_string)
+      @ascii_rabbit = {:name => "rabbit", :legs => 4, :ascii_string => @ascii_string}
     end
-
-    it "marshals objects"
-      # @store.set(@utf8_key, @ascii_rabbit)
-      # @store.get(@utf8_key).should eql(@ascii_rabbit)
 
     it "gets and sets raw values" do
       @store.set(@utf8_key, @ascii_string, :raw => true)
       @store.get(@utf8_key, :raw => true).bytes.to_a.should eql(@ascii_string.bytes.to_a)
     end
 
-    it "marshals objects on setnx"
-      # @store.del(@utf8_key)
-      # @store.setnx(@utf8_key, @ascii_rabbit)
-      # @store.get(@utf8_key).should eql(@ascii_rabbit)
+    it "marshals objects on setnx" do
+      @store.del(@utf8_key)
+      @store.setnx(@utf8_key, @ascii_rabbit)
+      retrievied_ascii_rabbit = @store.get(@utf8_key)
+      retrievied_ascii_rabbit.except(:ascii_string).should eql(@ascii_rabbit.except(:ascii_string))
+      JSON.load(JSON.generate(retrievied_ascii_rabbit[:ascii_string])).should eql(@ascii_string)
+    end
 
     it "gets and sets raw values on setnx" do
       @store.del(@utf8_key)
