+++ spec/redis/store/strategy/marshal_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe "Redis::Store::Strategy::Marshal" do
-  def setup
+  before(:each) do
     @store = Redis::Store.new :strategy => :marshal
     @rabbit = OpenStruct.new :name => "bunny"
     @white_rabbit = OpenStruct.new :color => "white"
@@ -9,83 +9,76 @@ describe "Redis::Store::Strategy::Marshal" do
     @store.del "rabbit2"
   end
 
-  def teardown
+  after :each do
     @store.quit
   end
 
   it "unmarshals on get" do
-    @store.get("rabbit").must_equal(@rabbit)
+    @store.get("rabbit").should eql(@rabbit)
   end
 
   it "marshals on set" do
     @store.set "rabbit", @white_rabbit
-    @store.get("rabbit").must_equal(@white_rabbit)
+    @store.get("rabbit").should eql(@white_rabbit)
   end
 
   if RUBY_VERSION.match /1\.9/
     it "doesn't unmarshal on get if raw option is true" do
-      @store.get("rabbit", :raw => true).must_equal("\x04\bU:\x0FOpenStruct{\x06:\tnameI\"\nbunny\x06:\x06EF")
+      @store.get("rabbit", :raw => true).should eql("\x04\bU:\x0FOpenStruct{\x06:\tnameI\"\nbunny\x06:\x06EF")
     end
   else
     it "doesn't unmarshal on get if raw option is true" do
-      @store.get("rabbit", :raw => true).must_equal("\004\bU:\017OpenStruct{\006:\tname\"\nbunny")
+      @store.get("rabbit", :raw => true).should eql("\004\bU:\017OpenStruct{\006:\tname\"\nbunny")
     end
   end
 
   it "doesn't marshal on set if raw option is true" do
     @store.set "rabbit", @white_rabbit, :raw => true
-    @store.get("rabbit", :raw => true).must_equal(%(#<OpenStruct color="white">))
+    @store.get("rabbit", :raw => true).should eql(%(#<OpenStruct color="white">))
   end
 
   it "doesn't unmarshal if get returns an empty string" do
     @store.set "empty_string", ""
-    @store.get("empty_string").must_equal("")
+    @store.get("empty_string").should eql("")
     # TODO use a meaningful Exception
-    # lambda { @store.get("empty_string").must_equal("") }.wont_raise Exception
+    # lambda { @store.get("empty_string").should eql("") }.wont_raise Exception
   end
 
   it "doesn't set an object if already exist" do
     @store.setnx "rabbit", @white_rabbit
-    @store.get("rabbit").must_equal(@rabbit)
+    @store.get("rabbit").should eql(@rabbit)
   end
 
   it "marshals on set unless exists" do
     @store.setnx "rabbit2", @white_rabbit
-    @store.get("rabbit2").must_equal(@white_rabbit)
+    @store.get("rabbit2").should eql(@white_rabbit)
   end
 
   it "doesn't marshal on set unless exists if raw option is true" do
     @store.setnx "rabbit2", @white_rabbit, :raw => true
-    @store.get("rabbit2", :raw => true).must_equal(%(#<OpenStruct color="white">))
-  end
-
-  it "marshals on set expire" do
-    @store.setex "rabbit2", 1, @white_rabbit
-    @store.get("rabbit2").must_equal(@white_rabbit)
-    sleep 2
-    @store.get("rabbit2").must_be_nil
+    @store.get("rabbit2", :raw => true).should eql(%(#<OpenStruct color="white">))
   end
 
   it "doesn't unmarshal on multi get" do
     @store.set "rabbit2", @white_rabbit
     rabbit, rabbit2 = @store.mget "rabbit", "rabbit2"
-    rabbit.must_equal(@rabbit)
-    rabbit2.must_equal(@white_rabbit)
+    rabbit.should eql(@rabbit)
+    rabbit2.should eql(@white_rabbit)
   end
 
   if RUBY_VERSION.match /1\.9/
     it "doesn't unmarshal on multi get if raw option is true" do
       @store.set "rabbit2", @white_rabbit
       rabbit, rabbit2 = @store.mget "rabbit", "rabbit2", :raw => true
-      rabbit.must_equal("\x04\bU:\x0FOpenStruct{\x06:\tnameI\"\nbunny\x06:\x06EF")
-      rabbit2.must_equal("\x04\bU:\x0FOpenStruct{\x06:\ncolorI\"\nwhite\x06:\x06EF")
+      rabbit.should eql("\x04\bU:\x0FOpenStruct{\x06:\tnameI\"\nbunny\x06:\x06EF")
+      rabbit2.should eql("\x04\bU:\x0FOpenStruct{\x06:\ncolorI\"\nwhite\x06:\x06EF")
     end
   else
     it "doesn't unmarshal on multi get if raw option is true" do
       @store.set "rabbit2", @white_rabbit
       rabbit, rabbit2 = @store.mget "rabbit", "rabbit2", :raw => true
-      rabbit.must_equal("\004\bU:\017OpenStruct{\006:\tname\"\nbunny")
-      rabbit2.must_equal("\004\bU:\017OpenStruct{\006:\ncolor\"\nwhite")
+      rabbit.should eql("\004\bU:\017OpenStruct{\006:\tname\"\nbunny")
+      rabbit2.should eql("\004\bU:\017OpenStruct{\006:\ncolor\"\nwhite")
     end
   end
 
@@ -98,24 +91,24 @@ describe "Redis::Store::Strategy::Marshal" do
 
     it "marshals objects" do
       @store.set(@utf8_key, @ascii_rabbit)
-      @store.get(@utf8_key).must_equal(@ascii_rabbit)
+      @store.get(@utf8_key).should eql(@ascii_rabbit)
     end
 
     it "gets and sets raw values" do
       @store.set(@utf8_key, @ascii_string, :raw => true)
-      @store.get(@utf8_key, :raw => true).bytes.to_a.must_equal(@ascii_string.bytes.to_a)
+      @store.get(@utf8_key, :raw => true).bytes.to_a.should eql(@ascii_string.bytes.to_a)
     end
 
     it "marshals objects on setnx" do
       @store.del(@utf8_key)
       @store.setnx(@utf8_key, @ascii_rabbit)
-      @store.get(@utf8_key).must_equal(@ascii_rabbit)
+      @store.get(@utf8_key).should eql(@ascii_rabbit)
     end
 
     it "gets and sets raw values on setnx" do
       @store.del(@utf8_key)
       @store.setnx(@utf8_key, @ascii_string, :raw => true)
-      @store.get(@utf8_key, :raw => true).bytes.to_a.must_equal(@ascii_string.bytes.to_a)
+      @store.get(@utf8_key, :raw => true).bytes.to_a.should eql(@ascii_string.bytes.to_a)
     end
   end if defined?(Encoding)
 end
